Fixed: OrderBook Time should be int64 not float64

+++ market/order_book.go
@@ -8,7 +8,7 @@ import (
 
 // OrderBook - A record of active buy and sell orders in a single market
 type OrderBook struct {
-	Time float64
+	Time int64
 	Asks []OrderBookRow `json:"asks"` // ordered from lowest to highest
 	Bids []OrderBookRow `json:"bids"` // ordered from highest to lowest
 }
